Merged eb/binstats -r3848:3906 into trunk.  These changes (1) fix a problem with gr_feval* where when called from a live flowgraph, they resulted in the call back into Python occuring without holding the Global Interpreter Lock causing a SIGSEGV.
(2) add gr_bin_statistics_f which combines statistics gathering
with a control state machine that allows spectrum sensing or related
applications that need to step through the spectrum to be built.

(3) usrp_spectrum_sense.py which ties all this together in an
application which steps through the spectrum a chunk at a time and
gathers statistics.  In the current version, the stats are gathered by
nothing is done with them.  Think of this as the framework for a real
application.  This code may require tuning of the --tune-delay and
--dwell-delay timeouts to ensure that the samples being processed are
associated with the given center frequency.


git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@3907 221aa14e-8319-0410-a670-987f0aec2ac5