SWIG compilation speedup!  Fixes ticket:9. Merged -r4203:4254 from eb/swig-split into trunk.
This refactors gnuradio_swig_python.{cc,py} into 5 separate .so's
These correspond to the runtime, general, filter and io directories,
and also includes a new directory, gengen.  gengen contains that part
of general that was machine generated.  This split is arbitrary, but
was useful for getting size of the swig generated glue code for
general down to about 2MB.

In addition, the swig glue is now compiled with -g1 -O1 instead of 
-g -O2.  With this change all the swig code now compiles in about 60%
of the time that it used to take.


git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@4255 221aa14e-8319-0410-a670-987f0aec2ac5