Doc fixes.  Merged eb/t367 -r11020:11025 to trunk.  This changeset moves the primary doxygen configuration under the top-level docs directory.  It creates a new "docs" top-level component that can be enabled/disabled using the configure --disable-docs option.  At this time, the --enable-doxygen option is still required to enable the generation of the doxygen documents.  I think the flag should probably be removed, and default to "yes" if we find doxygen on the path.  The user can disable the doc generation using --disable-docs if desired.
The doxygen config file has been modified such that doxygen is now
only run on the C++ sources.  No attempt is made to process the python
code using doxygen.  This results in a less confusing set of docs for
the the C++ API.  Straightening out the python docs is left for later.

Many classes are currently misclassified (\ingroup <wrong>).  That
will probably require another day of work, that I can't get to right now.



git-svn-id: http://gnuradio.org/svn/gnuradio/trunk@11027 221aa14e-8319-0410-a670-987f0aec2ac5