Evaluation fix in param.to_code().
Dont force an evaluation in to code unless the type is string or list.
Not doing so forces the variables to call evaluate before the namespace was bootstrapped.

This fixes a bug that came up when the validate was replaced with rewrite in flowgraph.import_data().
By replacing the validate, evaluate was only called once, and the namespace was not bootstrapped.