Flush pending errors in gr_udp_sink on disconnect()
On some systems (e.g., Debian/lenny) UDP errors are reported on the
following send() or recv() call.  To avoid having errors (such as
ECONNREFUSED) from an old connection showing up on the first write
to a new connection, we do a recv() on disconnect() to flush them.
This may not work for all errors on all systems, but it works in
some simple cases of interest.