Clean up lib/swig/Makefile.am, Makefile.common and Makefile.swig
Confirmed that it builds and make checks on all four combintations
of --{enable,disable}-{python,guile}.

Have not tested make dist, but expect that there may be some problems
with it.  I'm pretty sure that not all files that need to be removed
from the distribution are removed, and make clean may still be leaving
some files around.