docs: autotools Makefile.am generates empty swig_docs .i files; builds in autotools and cmake.
The autotools build does not auto-generate the Python documentation for each block like we do with the cmake build, but the empty .i files are required to build things regardless. The swig_doc.py and doxyxml files are distributed with the autotools build in case someone wants to generate this on their own (see README.doxyxml for how to do this).

Only for gnuradio-core and gr-digital currently.