pfb: bug fix for ccf version of the PFB clock sync to handle the d_osps parameter better.
The block now holds state between calls to work for the number of output symbols instead of requiring a certain number of available outputs (that may never be satisfied).

This commit also changes the number of samples needed in the QA code for the constellation receiver for a faster test.