uhd: fix dc offset call in uhd source
SWIG/python cannot differentiate the calls set_dc_offset
with an complex number or a boolean paramater.
It seems to take the boolean as a complex number 1 or 0.

Since the overloading is not working right:
renamed set_dc_offset to set_auto_dc_offset
for the boolean enable/disable paramater