filter: Fixing the Hilbert transform and QA.
The hilbert_fc block now takes the type of window as a parameter. Defaults to Hamming to a) maximize sidelobe suppression and b) that's how it's always behaved.

The QA codes were fixed to create Hilbert transform taps using Hamming window to account for the previous commits change to the rectangular window.