Add a QA_PYTHON_EXECUTABLE variable so QA code knows what python to use when cross compiling.
With this (and the prior) patch, you can mount the compile directory on the target and run
the QA code (as long as the paths are the same). The qtgui tests fail if there is no X
server. Volk seems to fail for hard float builds.

We'll need to look at adding thes changes to gr_modtool.

Signed-off-by: Philip Balister <philip@balister.org>