wxgui: Make sure to only start the flow graph once all init is done
Currently we start the flow graph in the stdframe constructor.
However at that point, the frame isn't attached to the app and it's not
"shown" yet which means some UI operations can still fail.

This is a race condition which happens on OSX. You can reproduce it on
linux as well by adding a sleep(1) after starting the flow graph.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>