controlport: moving the logger call that publishes Thrift's endpoint to i_impl().
Using to verify that the hostname and OS selected port number are
eventually determined correctly.

However, as i_imp() may be called many times throughout the lifetime
of a flowgraph, this can't be the final place where the endpoint is
generated. It appears that the initial calls within the first few
milliseconds of TServerSocket creation may return an incorrect port
number when getPort() is queried (i.e. 0 if a randomly open / selected
port is requested vs. the selected port number).

Endpoint string generation and the call to the logger should probably
be moved to a function that executes only once after the results from
getPort() have had time to settle.