analog: Power Squelch now emit tags on start/end of bursts
The block will emit a tag with the key pmt::intern("squelch_sob")
with the value of pmt::PMT_NIL on the first item it passes, and with
the key pmt::intern("squelch:eob") on the last item it passes.

Added this to the `make` documentation.

Also: made the GRC XML's defaults reflect the C++ default args.

Also: indentation was mixed; unified the paragraphs I worked on.