runtime: fixes a problem with tag pruning.
Looked like a race condition if two blocks are reading from the same
buffer; one block calls prune before the other one can, and then it no
longer gets its tags. This puts back in the original protections we
used when pruning with vectors, only we don't have to restart the
iterator at the beginning and can exit sooner because of the sort
order. Likely slower than the multimap erase, so we will have to
reinvestigate what kind of performance hit we take and if we can do
better.