docs: Reviving Python manual.
I'd like to only produce a single manual, and the Sphinx docs take too
much manual labor to keep up-to-date. Still, there is a lot of useful
tools and classes only in Python that need to be exposed in a manual,
so this is the quickest way to get there.

This flattens the sphinx documentation structure so each component has
a <component>_blocks.rst where the actual GNU Radio blocks are managed
and a <component>.rst for any other non-block or Python-only code
(hier_block2, functions, and classes).

The new flattened file structure should make it easier to update, but
the index.rst should have some work done to make structure information
better.