Fix the VOLKized fft_vcc_fftw and QA code
The QA code did not cover all the cases being VOLKized, and thus failed to notice the incorrect pointer math.
This QA code (correctly!) fails on the previous commit, and (correctly!) passes with this fix to fft_vcc_fftw.cc.
Thanks to awalls-cx18 for noticing this, and providing the correct array indexing call to VOLK.
Fixes issue #844