gr-utils: Update error message
Prevent people from running into the problem that "import pylab" doesn't work because python tkinter is not installed.
For example on a fresh Ubuntu or Mint system it is not enough to install python-matplotlib.
However, the true error message:
"ImportError: No module named _tkinter, please install the python-tk package"
is obfuscated towards the user.

Another good option would be to not catch this import error but just show the true error message from python to the user.