digital: addresses issue #876.
The constellation receiver controlport interface was modeled after the
costas_loop_cc block, but the inheritance of the two blocks to
control_loop was different. This changes the inheritance of
constellation_reciever_cb.

It also fixes the name of the "loop_bw" parameter to be consistent
between the set and get ControlPort interfaces.