gr-fec: Removing 2 bad matrices from the LDPC alist file collection
First: if trying to use n_2400_k_1198_gap_33.alist, the following error will be returned:

Error in set_parameters_for_encoding while finding inverse_phi: Error in calc_inverse_mod2(): The matrix inverse found is not valid.
Tip: verify that the correct gap is being specified for this alist file.
Traceback (most recent call last):
.
.
.
    return _fec_swig.ldpc_H_matrix_make(*args, **kwargs)
RuntimeError: set_parameters_for_encoding

What's happening is that when this matrix is read in, various submatrices are set to be used later for encoding. Then the inv(phi) matrix is calculated. This matrix must be non-singular so that the inv can be found - it's a requirement for this particular matrix format. Anyway, the inverse found is invalid, for some reason. I tried to perform some matrix manipulations to get it to a usable state, where the phi matrix is not nonsingular, but no joy. So, this matrix isn't usable with the encoder as-is, and should come out of the matrix collection.

-----

Second: I think the matrix n_2400_k_1198_gen_matrix.alist was intended to be the corresponding generator matrix, in systematic form, for the matrix mentioned above. However, it's the wrong size to be so - it's 1198 x 2400, which is the same size as the matrix in n_2400_k_1198_gap_33.alist, and that's wrong. I suspect that this matrix was created using python function getSystematicGmatrix, which was buggy to use for this purpose (I have just submitted some fixes  to those python functions).