qtgui: Adds a QTGUI edit box block that emits a message when changed.
Unlike the current QT GUI Entry Widget, this is an actual GNU Radio
block with an output message port 'msg'. When the editing is finished
(enter or change in scope), it triggers the editing_finished method
that publishes the text in the box as a message.

Also allows setting of the value through an input message port.

Adds example test_qtgui_msg.grc to explore this.