qtgui: raised time sink inputs to 12/24 and added GRC check for 5/10
This commit raises the number of inputs to gr::qtgui:time_sink_c and f
to 12 and 24, respectively. However, the GRC XML wrapper for the block
has a hard-coded limit of 5/10; so a check was added to ensure the
input field in GRC is restricted to this amount. The future fix should
be to rewrite the XML (or YAML in 3.8) to operate as a for loop.