Removing building-boost README: no longer necessary, out of place
With 3.8, we'll no longer support distros that ship a Boost too old for
GNU Radio. (And even on old distros, there's often an official way of
installing Boost 1.48 or later parallely without messing up the whole
system.)

Even if that changes in the future, I really don't think building
third-party dependencies is something that should be documented within
the GNU Radio source tree.