python gateway: permit a variable number of ports
C++ block in/out ports are described by io_signature, which holds
the min number of ports, max number of ports, and a list of type
lengths. If len(type lengths) < max number of ports, the last type
is assumed to apply to the remaining ports. When a flowgraph is built,
a variable number of ports can be connected.

Python block ports are specified using a list of numpy types. The
number of ports is fixed in the block specification, and the flowgraph
must connect exactly that number of ports.

This patch changes the Python behavior (while maintaining backward
compatibility) to match the C++ behavior. A new py_io_signature
class is used to specify port number limits, along with a list of
numpy types, like this:

  in_sig = gr.py_io_signature(2, -1, ['float32', 'float32'])

where in_sig is passed to the block super's __init__ as before.