blocks: Convert GRC examples to YAML format
This commit converts the examples in `gr-blocks` to the new YAML format.
This conversion is tracked in #2285.

The assertion in `grc/blocks_multiply_matrix_xx.block.yml` to check if
the selected tag propagation policy and the given matrix are compatible
is disabled. This might be an issue with `dtype: enum` and its
resolution in Mako.