grc: fix generation of hier2 block in case of multiple message ports
If a hier2 block contains more than one ( input or output ) message port
GRC generates a wrong yml description of this block.
If you restart GRC, GRC terminates with an error messages is no longer usable.

This can be verified with the example here: "gr-blocks/examples/msg_passing/hier"
This example is running in 3.7 without any problems.

The modification in platform.py avoids the termination of grc
if there is an error in the flowgraph to be loaded.