cmake: Add `-undefined dynamic_lookup` linker flag back for macOS
For Homebrew, gnuradio is built against brewed Python 3.8. This means
that GR will link against that Python.

If later on the user has another Python 3 in their path, running `gnuradio`
will fail with the `Abort Trap 6` error message.

On Linux, you do not need this flag, as this is the default behaviour of
the compiler. On Mac, you need to explicitly ask for dynamic lookup.

This setting was removed in commit ab2fb35677e38a384df3f9503d1f45f64bbc0374
and this patch adds it back in.