fft: window: Provide default value for beta param on window::build
The API call fft::window::build() takes a third parameter, beta, that
only applies for the Kaiser window. For other windows, it is necessary
to provide a dummy value to call this function.

The declared-deprecated version firdes::window() on the other hand does
not require specifying beta when not needed. Instead, we provide
a default value of 6.76 which will still generate a valid (and generally
useful) Kaiser window, but means we don't have to specify beta for other
windows.