uhd: uhd_fft: Add support for --power argument
This enables absolute power support for uhd_fft. For example:

    uhd_fft --power -20 --args [...]

will run uhd_fft with all the usual settings, but will set the gain such
that the reference power level is -20 dBm. The y-axis will be labeled
differently ("RX Power Level (dBm)") and the gain slider will be
replaced by a power reference level slider.

Reading the actual received power from the display requires some DSP
knowledge. First, the choice of window will move energy into sidelobes,
and of course, scalloping loss plays a role. By choosing a boxcar window,
and moving the signal close to the center of a bin, it is possible to
measure the power of a single tone, assuming the USRP has been
previously calibrated.

A more accurate y-label would be "dBm per FFT bin after FFT processing"
but that is considered a bit unwieldy.

The y-axis scaling is achieved by throwing in a multiplier that moves
a full-scale signal to the corresponding power value, and using the
recently introduced features to normalize windows. That means scaling
the reference power will be visible in the scope plot, but not in the
FFT plot unless the signal starts to clip.