qtgui:Added Time Raster Sink Control Axis Configuration
This update allows a time raster sink's axis minimum, maximum,
and label to now be specified in the block configuration options.
This allows the block to be formatted to match a waterfall plot
in terms of having frequency on the lower axis.

Based on PR from ghostop14 <ghostop14@gmail.com>

Signed-off-by: Jeff Long <willcode4@gmail.com>