modtool: Fix `rm`, `disable`, `rename` and `makeyml` similar naming
* modtool: Fix issue #2672 for `rm`

In the event that the user specifies only one specific block to delete,
this PR ensures that the files for said block is not confused with files
of similarly named blocks.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>

* modtool: Fix issue #2672 for `rename` and more

This commit fixes the following for the `rename` command.
1. Issue #2672 that also affects `rename`
2. Fixes errorneous replacement of similarly named blocks in CMakeLists
3. Replaced the file search in `_run_file_rename` from `file.find` to
`re.search` so that block names will not be confused with names with one
or two prefix characters (e.g. "decoder" vs "adecoder").
4. Renames the bind function in `python_bindings.cc` as well.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>

* modtool: Fix Issue #2672 for `disable` & more

When the user specifies a block name, only the block with that specific
blockname should be disabled. Prior to this PR, all blocks matching said
specified block name are disabled. This PR fixes it.

This PR also fixes what appears to be a spelling error `self._info`. The
error is thrown for OOT module version 38:
Attribute error `'ModToolDisable' object has no attribute '_info'`

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>

* modtool: Fix Issue #2672 for `makeyaml`

This PR fixes issue #2672 which also affects the `makeyaml` command.

Signed-off-by: Solomon Tan <solomonbstoner@yahoo.com.au>