gr-iio: pluto source /sink fails to set paramter if german locale is used
pluto source / sink fails to set hardwaregain with

gr::log :WARN: iio::device::set_params - Unable to write attribute in_voltage0_hardwaregain: -22
gr::log :WARN: iio::device::set_params - Unable to write attribute out_voltage0_hardwaregain: -22

gain values are treated as double. The conversion std::to_string() uses the locale setting.
So in the US std::to_string(10.) leads to "10.0" while in Germany the result is "10,0" which the pluto handles as error.
This problem only exists if real values are converted. So check if the converted string contains a , as separator and change it to a dot .

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>