blocks: Add symbolinterleaver block
* blocks: Add symbolinterleaver block

This commit adds a symbol interleaver to gr-blocks.
The idea is to specify an interleaver sequence to be able to interleave
and deinterleave vectors of constant size. For convenience, the
deinterleaver sequence is always derived. Thus, it is not necessary to
manually derive the corresponding deinterleaver sequence.

See issue #3637 for a preliminary discussion.
https://github.com/gnuradio/gnuradio/issues/3637

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Add type specializations to symbolinterleaver

All common GNU Radio port types are supported now, i.e. bb, cc, ff, ii,
ss. More type definitions may be added easily.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Add GRC definition for symbolinterleaver

In order to use the symbolinterleaver in GRC, the appropriate block
definition was added.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Fix fmtlib include for conda build

The conda CI fails to build GR with `#include <fmt/core.h>`. According
to online discussions, the correct include is `#include
<spdlog/fmt/fmt.h>`. There's some hope that this change makes all CI
pass.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Rename to blockinterleaver

The blockinterleaver was introduced as symbolinterleaver. Following a
suggestion to this PR. The block is now called blockinterleaver.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Fix MSVC non-standard keyword handling

Since MSVC doesn't conform to logical keywords `and` and `or` by
default, we can't use them. The `/permissive-` flag might change that
but breaks the build in other places.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Add huge interleaver test

Since the `blockinterleaver` calls `set_output_multiple` with
`interleaver_indices.size()`, we need to make sure this configuration
doesn't break things.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Fix for missing keyword

`not` is a C++ keyword that is not available by default in some
compilers. Albeit, it is a part of the C++ standard.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Add block interleaver example

This new block lacked an example. In the process some issues where
identified and fixed.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>

* blocks: Move blockinterleaving to kernel namespace

The actual interleaver implementation is a kernel and thus should reside
in the kernel namespace.

Signed-off-by: Johannes Demel <demel@ant.uni-bremen.de>