uhd: rfnoc: Bind set/get_property APIs
This adds gr::uhd::rfnoc_block::set_property and get_property to the
Python API.

The tricky bit is that these are templated functions. The Python
versions therefore receive an optional, fourth argument 'type_hint'
which loosely corresponds to the template argument in C++.

If there is no type hint, we try and guess the correct type based on
what came in from Python. However, that can be ambiguous (e.g., Python's
integer could be an int, size_t, long long, uint32_t or whatever in
C++). For these cases, we allow a string-based representation of the
type.

Signed-off-by: Martin Braun <martin.braun@ettus.com>