fec: fix some bugs and refactor async_decoder
This fixes a few bugs and shortcomings in async_decoder and refactors
it to reduce code duplication by merging the decode_unpacked() and
decoded_packed() functions into a single decode() function.

The problems solved are the following:

- decode_packed() wasn't considering the 'diff' (Viterbi tail) properly,
  unlike decode_unpacked(). This caused undefined behaviour when working
  in packed output mode with the Viterbi decoder in terminated mode.
  Occasionally this would cause bit errors in the last bits of a packet.
  See #6445 for a very similar problem affecting tagged_decoder.

- decode_packed() didn't support multiple blocks for fixed-size decoders
  (unlike decode_unpacked()).

- In both decode_unpacked() and decode_packed(), the input conversion
  to uint8_t could overflow (see #6433). The behaviour has been changed
  to clipping, which is less harmful.

Taking into account that some of these problems affected decode_packed() and not
decode_unpacked() because the two functions were quite similar but had some
differences, it seems a good idea to reduce the code duplication by merging the
two functions into one.

Additionally, there are these improvements:

- The opportunity to use volk_32f_s32f_x2_convert_8u (when it is released)
  for the input conversion to uint8_t has been identified in a comment.

- When there is a decoder shift and no input conversion, the shift is
  applied with volk_32f_s32f_add_32f rather than with an
  element-by-element loop.

Signed-off-by: Daniel Estévez <daniel@destevez.net>