grc: Fix way-over-backslashing in file paths
Each time File -> Open Recent is used on Windows, the number of
backslashes in the filename doubles. This happens because a GLib.Variant
object is passed through str(), which produces a quoted (and escaped)
string. The quotes are then removed, but the escaping remains. The
underlying string can instead be accessed using the get_string() method.

See https://xkcd.com/1638/ for further details. ;)

Signed-off-by: Clayton Smith <argilo@gmail.com>